requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2
faker
